Refactor code into more files and start adapting code to the changes
